# server/openapi.yaml
openapi: 3.0.0
info:
  title: StellarYield API
  description: Backend API for the StellarYield yield farming and portfolio management platform
  version: 1.0.0
  contact:
    name: StellarYield Team
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.stellaryield.com
    description: Production server
...
paths:
...
  /api/yields:
    get:
      summary: Get current yield data
      description: Returns current yield information for all available vaults with APY, TVL, and risk ratings
      tags:
        - Yields
      responses:
        '200':
          description: Successful response with yields data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/YieldData'
        '500':
          description: Failed to fetch yield data
...
  /api/leaderboard:
    get:
      summary: Get user leaderboard
      description: Returns ranked list of users by TVL, yield earned, or other metrics
      tags:
        - Leaderboard
      parameters:
        - name: metric
          in: query
          schema:
            type: string
            enum: [tvl, yield, roi]
            default: tvl
          description: Metric to rank by
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [week, month, all]
            default: month
          description: Time period for ranking
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
          description: Number of results to return
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
          description: Pagination offset
      responses:
        '200':
          description: Leaderboard data retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      $ref: '#/components/schemas/LeaderboardEntry'
                  total:
                    type: integer
        '400':
          description: Invalid parameters
...
  /api/alerts:
    post:
      summary: Create APY alert
      description: Creates a new APY alert for a wallet and vault
      tags:
        - Alerts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: Stellar wallet address (G...)
                vaultId:
                  type: string
                  description: Vault identifier
                condition:
                  type: string
                  enum: [above, below]
                  description: Alert condition
                thresholdValue:
                  type: number
                  description: APY threshold value (0-1000)
                email:
                  type: string
                  format: email
                  description: Email address for notifications
              required:
                - walletAddress
                - vaultId
                - condition
                - thresholdValue
                - email
      responses:
        '201':
          description: Alert created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          description: Invalid alert parameters
        '429':
          description: Maximum alerts per user exceeded
        '500':
          description: Failed to create alert
...
  /api/auth/challenge:
    post:
      summary: Request authentication challenge
      description: Generates a challenge for wallet signing
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                walletAddress:
                  type: string
                  description: User's Stellar wallet address
      responses:
        '200':
          description: Challenge generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  challenge:
                    type: string
                  walletAddressType:
                    type: string
                    enum: [account, contract]

    /metrics:
      get:
        summary: Prometheus scrape endpoint
        description: Prometheus-compatible metrics exposition. Returns plain-text metrics. This endpoint is protected by `METRICS_TOKEN` in production; unauthenticated requests return 404 to avoid token discovery.
        tags:
          - Metrics
        responses:
          '200':
            description: Prometheus metrics in text exposition format
            content:
              text/plain:
                schema:
                  type: string
          '404':
            description: Metrics endpoint not available or unauthorized
          '429':
            description: Too many requests (rate limited)
          '500':
            description: Internal server error while collecting metrics

    /api/audit-monitoring/alerts:
      get:
        summary: Get recent audit alerts
        description: Returns recent audit/monitoring alerts. Requires admin authentication.
        tags:
          - AuditMonitoring
        parameters:
          - name: severity
            in: query
            schema:
              type: string
              enum: [critical, warning, info]
          - name: type
            in: query
            schema:
              type: string
          - name: limit
            in: query
            schema:
              type: integer
              default: 100
        responses:
          '200':
            description: Alerts retrieved successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    count:
                      type: integer
                    alerts:
                      type: array
                      items:
                        $ref: '#/components/schemas/AuditAlert'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve alerts

    /api/audit-monitoring/alerts/{severity}:
      get:
        summary: Get alerts filtered by severity
        description: Returns recent audit alerts for the given severity. Requires admin authentication.
        tags:
          - AuditMonitoring
        parameters:
          - name: severity
            in: path
            required: true
            schema:
              type: string
          - name: limit
            in: query
            schema:
              type: integer
              default: 100
        responses:
          '200':
            description: Alerts retrieved successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    count:
                      type: integer
                    severity:
                      type: string
                    alerts:
                      type: array
                      items:
                        $ref: '#/components/schemas/AuditAlert'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve alerts

    /api/audit-monitoring/status:
      get:
        summary: Get monitoring status
        description: Returns current status of audit monitoring subsystem. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring status retrieved
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to retrieve monitoring status

    /api/audit-monitoring/report:
      get:
        summary: Generate a monitoring report
        description: Generates and returns a monitoring report (async generation may be implemented). Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring report generated
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/MonitoringReport'
          '403':
            description: Admin access required
          '500':
            description: Failed to generate report

    /api/audit-monitoring/export:
      get:
        summary: Export monitoring data
        description: Returns a JSON export of monitoring data as an attachment. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring export (application/json attachment)
            content:
              application/json:
                schema:
                  type: object
          '403':
            description: Admin access required
          '500':
            description: Failed to export monitoring data

    /api/audit-monitoring/start:
      post:
        summary: Start audit monitoring
        description: Starts audit monitoring with provided configuration. Requires admin authentication.
        tags:
          - AuditMonitoring
        requestBody:
          required: false
          content:
            application/json:
              schema:
                type: object
        responses:
          '200':
            description: Monitoring started
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    status:
                      $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to start monitoring

    /api/audit-monitoring/stop:
      post:
        summary: Stop audit monitoring
        description: Stops audit monitoring. Requires admin authentication.
        tags:
          - AuditMonitoring
        responses:
          '200':
            description: Monitoring stopped
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    success:
                      type: boolean
                    message:
                      type: string
                    status:
                      $ref: '#/components/schemas/MonitoringStatus'
          '403':
            description: Admin access required
          '500':
            description: Failed to stop monitoring

  # ── Deposit Simulation ──────────────────────────────────────────────────────
...
  /api/relayer/fee-bump:
    post:
      summary: Sign a fee-bump transaction
      description: Signs a Stellar fee-bump transaction through the configured relayer.
      tags:
        - Relayer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Fee-bump transaction signed
        '429':
          description: Rate limit exceeded
  /api/relayer/status:
    get:
      summary: Get relayer health status
      description: Returns bridge relayer health metrics including queue depth, replay protection state, and recent activity.
      tags:
        - Relayer
      responses:
        '200':
          description: Relayer status retrieved
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Failed to retrieve relayer status
...
  /api/events:
    get:
      summary: Get recent events
      description: Returns the latest persisted events when the Prisma client and database are available.
      tags:
        - Events
      responses:
        '200':
          description: Recent events returned successfully
        '503':
          description: Events database is unavailable
  /api/recommend:
    post:
      summary: Generate a yield recommendation
      description: Produces a recommendation from user preferences, risk tolerance, APY, liquidity, and volatility inputs.
      tags:
        - Recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Recommendation generated successfully
  /api/recommend/timeline:
    get:
      summary: Get recommendation timeline
      description: Returns recorded recommendation timeline entries for a user.
      tags:
        - Recommendations
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Recommendation timeline returned successfully
  /api/stress-scenarios/run:
    post:
      summary: Run a stress scenario
      description: Runs an APY collapse, liquidity drain, or oracle shock stress scenario using request inputs.
      tags:
        - Risk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scenario:
                  type: string
                  enum: [apy-collapse, liquidity-drain, oracle-shock]
      responses:
        '200':
          description: Stress scenario result returned successfully
        '400':
          description: Invalid scenario
  /api/auth/verify:
    post:
      summary: Verify a Stellar auth challenge
      description: Verifies a signed Stellar authentication challenge request.
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Auth challenge verified
        '400':
          description: Invalid auth verification request
...
  /api/backtest:
    get:
      summary: Validate a backtest request
      description: Validates a backtest request payload assembled from query parameters.
      tags:
        - Backtest
      responses:
        '200':
          description: Backtest request validated
        '400':
          description: Backtest request failed validation
...
  /api/incidents:
    get:
      summary: List incidents
      description: Returns a paginated list of protocol incidents with optional filters.
      tags:
        - Incidents
      responses:
        '200':
          description: Incidents returned successfully
    post:
      summary: Create an incident
      description: Creates a protocol incident record.
      tags:
        - Incidents
      responses:
        '201':
          description: Incident created
        '400':
          description: Missing required fields
...
  /api/presets:
    get:
      summary: List allocation presets
      tags:
        - Presets
      parameters:
        - name: wallet
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Presets returned successfully
        '400':
          description: Wallet query parameter is required
    post:
      summary: Create an allocation preset
      tags:
        - Presets
      responses:
        '201':
          description: Preset created
        '400':
          description: Invalid preset input
  /api/fees:
    get:
      summary: Get fee oracle estimate
      tags:
        - Fees
      responses:
        '200':
          description: Fee estimate returned successfully
        '500':
          description: Unable to estimate fees
...
  /api/admin:
    get:
      summary: Get admin status
      description: Representative admin route prefix; admin-only operations require privileged access.
      tags:
        - Admin
      responses:
        '200':
          description: Admin data returned successfully
        '403':
          description: Admin access required
...
  /api/weekly-reports:
    get:
      summary: List weekly yield reports
      tags:
        - WeeklyReports
      responses:
        '200':
          description: Weekly reports returned successfully
        '403':
          description: Admin access required
  /api/correlation:
    get:
      summary: Get correlation matrix
      tags:
        - Correlation
      parameters:
        - name: window
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 365
      responses:
        '200':
          description: Correlation matrix returned successfully
        '400':
          description: Invalid window parameter
...
  /api/contacts:
    get:
      summary: List encrypted contacts
      description: Returns encrypted contacts for the authenticated wallet.
      tags:
        - Contacts
      responses:
        '200':
          description: Contacts returned successfully
    post:
      summary: Create encrypted contact
      description: Stores a new encrypted contact for the authenticated wallet.
      tags:
        - Contacts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                encryptedName:
                  type: string
                encryptedAddress:
                  type: string
      responses:
        '201':
          description: Contact created
    delete:
      summary: Delete all contacts
      description: Removes all contacts for the authenticated wallet.
      tags:
        - Contacts
      responses:
        '204':
          description: Contacts deleted
...
components:
  schemas:
...
    YieldData:
      type: object
      properties:
        protocol:
          type: string
          example: Blend
        asset:
          type: string
          example: USDC
        apy:
          type: number
          format: double
          example: 6.5
        tvl:
          type: number
          format: double
          example: 12000000
        risk:
          type: string
          enum: [Low, Medium, High]
        fetchedAt:
          type: string
          format: date-time
...
    LeaderboardEntry:
      type: object
      properties:
        rank:
          type: integer
        walletAddress:
          type: string
        tvl:
          type: number
          format: double
        yieldEarned:
          type: number
          format: double
        roi:
          type: number
          format: double
...
    Alert:
      type: object
      properties:
        id:
          type: string
        walletAddress:
          type: string
        vaultId:
          type: string
        condition:
          type: string
          enum: [above, below]
        thresholdValue:
          type: number
        email:
          type: string
          format: email
        isActive:
          type: boolean
        triggeredAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    AuditAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        severity:
          type: string
          enum: [critical, warning, info]
        message:
          type: string
        metadata:
          type: object
        createdAt:
          type: string
          format: date-time
    MonitoringStatus:
      type: object
      properties:
        running:
          type: boolean
        startedAt:
          type: string
          format: date-time
          nullable: true
        lastCheckedAt:
          type: string
          format: date-time
          nullable: true
        alertCount:
          type: integer
        listeners:
          type: array
          items:
            type: string
    MonitoringReport:
      type: object
      properties:
        generatedAt:
          type: string
          format: date-time
        summary:
          type: object
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/AuditAlert'
        metrics:
          type: object
        exportUrl:
          type: string
          description: URL to download a full export if available (admin-only)
...